fix problems

+++ models.py
@@ -11,7 +11,7 @@ def init_db():
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             name TEXT NOT NULL,
             description TEXT,
-            price REAL NOT NULL,
+            price INTEGER NOT NULL,
             image_paths TEXT,
             category TEXT NOT NULL
         )
